+++ pz_be_services/main.py
@@ -1,12 +1,6 @@
 from fastapi import FastAPI
-from core.logger import get_logger
+from routers.v1.health_router import router
 
-app = FastAPI(title="Project Zero Backend Service")
-logger = get_logger()
-db_logger = get_logger("DB_Logger", True)
+app = FastAPI()
 
-@app.get("/health")
-def health_check():
-    logger.debug("Health check.")
-    db_logger.debug("DB Health check.")
-    return {"status": "ok"}
+app.include_router(router, prefix="/v1")
